+++ src/handlers/threadRowViewHandler.js
